deploy/aws/requirements.txt
boto3==1.26.160
python-dotenv==1.0.0